delete MyChat and OtherChat components, made some changes in Messages and Messages components
