+++ package.json
@@ -35,7 +35,7 @@
     "react-native-permissions": "^3.0.0",
     "react-native-safe-area-context": "^3.2.0",
     "react-native-sound": "^0.11.0",
-    "react-native-track-player": "^1.2.7",
+    "react-native-track-player": "^1.2.3",
     "react-native-vector-icons": "^8.1.0"
   },
   "devDependencies": {
